pyproject.toml
[tool.poetry]
name = "imagesteg"
version = "0.2.0"
description = "A simple image steganalysis (steganography analysis) tool in Python"
authors = ["nolze <nolze@int3.net>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/nolze/imagesteg"
homepage = "https://github.com/nolze/imagesteg"
packages = [{ include = "imagesteg" }]

[tool.poetry.scripts]
imagesteg = "imagesteg.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
bitstring = "^4.1.4"
bottle = "^0.12.25"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
